1.训练代码/YoloV5_PyQt5-main/YoloV5_PyQt5-main/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.0
decorator==5.1.1
einops==0.8.1
executing==2.0.1
fast-histogram==0.13
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.62.2
humanfriendly==10.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==7.0.0
importlib-resources @ file:///C:/b/abs_d0dmp77t95/croot/importlib_resources-suite_1704281892795/work
install==1.3.5
intel-openmp==2021.4.0
ipykernel==6.27.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.13.1
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib @ file:///C:/b/abs_085jhivdha/croot/matplotlib-suite_1693812524572/work
matplotlib-inline==0.1.6
mkl==2021.4.0
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
munkres==1.1.4
nest-asyncio==1.5.8
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.14.1
onnxoptimizer==0.2.7
onnxruntime==1.16.0
opencv-python==4.11.0.86
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.1.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==25.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pyDeprecate==0.3.1
Pygments==2.4.0
PyMySQL==1.1.1
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==1.5.0
pytz==2025.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
scipy==1.10.1
seaborn==0.13.2
sip @ file:///C:/b/abs_edevan3fce/croot/sip_1698675983372/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==1.12.0
torch-tools==0.1.5
torchaudio==0.12.0
torchmetrics==1.3.2
torchtext==0.13.0
torchvision==0.13.0
tornado @ file:///C:/b/abs_0cbrstidzg/croot/tornado_1696937003724/work
tqdm==4.67.1
traitlets==5.14.0
typing_extensions==4.11.0
tzdata==2025.1
ultralytics==8.3.80
ultralytics-thop==2.0.14
urllib3==2.2.1
visdom==0.2.4
wcwidth==0.2.12
websocket-client==1.8.0
Werkzeug==3.0.2
yarl==1.9.4
zipp==3.17.0
